Fix warning.
GitOrigin-RevId: 033adb463f56b8135180dd22d4c48e82b967fe39

+++ tddb/td/db/TQueue.cpp
@@ -223,7 +223,7 @@ class TQueueImpl : public TQueue {
     if (from_id.value() > q.tail_id.value() + 10) {
       return Status::Error("Specified from_id is in the future");
     }
-    if (from_id.value() < get_queue_head(q).value() - MAX_QUEUE_EVENTS) {
+    if (from_id.value() < get_queue_head(q).value() - static_cast<int32>(MAX_QUEUE_EVENTS)) {
       return Status::Error("Specified from_id is in the past");
     }
